rename sharing plugin source files and target
...use ghostcloud as a prefix
