+++ src/sharing/ghostcloudconsts.h
