+++ src/sharing/ghostcloudplugininfo.h
