+++ src/sharing/ghostcloudshareplugin.h
@@ -2,7 +2,7 @@
 #define NEXTCLOUDSHAREPLUGIN_H
 
 #include <TransferEngine-qt5/transferplugininterface.h>
-#include "nextcloudconsts.h"
+#include "ghostcloudconsts.h"
 #include <QObject>
 
 class Q_DECL_EXPORT NextcloudSharePlugin : public QObject, public TransferPluginInterface
